Introduce new exception type
