+++ waffle-chai/src/matchers/calledOnContract/error.ts
@@ -1,0 +1,5 @@
+export class ProviderWithHistoryExpected extends Error {
+  constructor() {
+    super('calledOnContract matcher requires provider that support call history');
+  }
+}
